Breach laser changes
